+++ contracts/TokensFlowTest.sol
@@ -8,7 +8,7 @@ import "./TokensFlow.sol";
 
 contract TokensFlowTest is TokensFlow
 {
-    uint timestamp = 10000; // an arbitrary value for testing
+    int timestamp = 10000; // an arbitrary value for testing
 
     function mint(address _to, uint256 _id, uint256 _value, bytes calldata _data) external {
         require(tokenOwners[_id] == msg.sender);
@@ -18,15 +18,15 @@ contract TokensFlowTest is TokensFlow
         _doMint(_to, _id, _value, _data);
     }
 
-    function _currentTime() internal override view returns(uint256) {
+    function _currentTime() internal override view returns(int) {
         return timestamp;
     }
 
-    function currentTime() external view returns(uint256) {
+    function currentTime() external view returns(int) {
         return _currentTime();
     }
 
-    function setCurrentTime(uint256 _timestamp) external {
+    function setCurrentTime(int _timestamp) external {
         require(_timestamp >= timestamp);
         timestamp = _timestamp;
     }
